+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next
 * Marstek B2500: Fix wrong gauge display for panel 2 and output 2
+* Marstek B2500: also display current battery change level in Wh, not only percentage
 
 #### 0.83.0
 * Add first start onboarding screen